% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{wtm_parse_json}
\alias{wtm_parse_json}
\title{Parse WTM targeting JSON data}
\usage{
wtm_parse_json(raw)
}
\arguments{
\item{raw}{raw JSON}
}
\value{
returns a tibble
}
\description{
Parse WTM targeting JSON data
}
