% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{wtm_parse_json2}
\alias{wtm_parse_json2}
\title{Parse WTM targeting JSON data into list columns}
\usage{
wtm_parse_json2(raw)
}
\arguments{
\item{raw}{raw JSON}
}
\value{
returns a tibble
}
\description{
Parse WTM targeting JSON data into list columns
}
