% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{wtm_parse_targeting}
\alias{wtm_parse_targeting}
\title{Parse WTM targeting data}
\usage{
wtm_parse_targeting(col, p)
}
\arguments{
\item{col}{column that includes raw JSON}

\item{p}{used for iterator for progressr}
}
\value{
returns a tibble
}
\description{
Parse WTM targeting data
}
